requirements.txt
Django==1.11.7
django-debug-toolbar==1.9.1
djangorestframework==3.7.3
et-xmlfile==1.0.1
jdcal==1.3
openpyxl==2.4.9
pytz==2017.3
sqlparse==0.2.4
